refactor: extract contracts runtime title normalization helper
