+++ src/components/contracts/AdminContractsWorkspace.tsx
@@ -1,8 +1,6 @@
 "use client";
-
 import Link from "next/link";
 import { useCallback, useEffect, useMemo, useState } from "react";
-
 import {
   adminContractsCopyByLocale,
   contractApprovalStatusLabelByLocale,
@@ -14,13 +12,13 @@ import {
 } from "@/components/contracts/copy";
 import { resolveContractDocumentActionRequest } from "@/components/contracts/action-payloads";
 import { normalizeContractsErrorMessageForRuntime, readJson } from "@/components/contracts/http";
+import { normalizeContractsEntityTitle } from "@/components/contracts/runtime-copy-helpers";
 import {
   type AdminContractDocument as ContractDocument,
   type ContractDocumentAction,
   type ContractTemplate
 } from "@/components/contracts/types";
 import { useI18n } from "@/lib/i18n/provider";
-
 export default function AdminContractsWorkspace() {
   const { locale } = useI18n();
   const isKoLocale = locale === "ko";
@@ -30,7 +28,6 @@ export default function AdminContractsWorkspace() {
   const templateStatusLabels = contractTemplateStatusLabelByLocale[locale];
   const documentStatusLabels = contractDocumentStatusLabelByLocale[locale];
   const approvalStatusLabels = contractApprovalStatusLabelByLocale[locale];
-
   const [templates, setTemplates] = useState<ContractTemplate[]>([]);
   const [documents, setDocuments] = useState<ContractDocument[]>([]);
   const [employeeId, setEmployeeId] = useState("");
@@ -58,25 +55,6 @@ export default function AdminContractsWorkspace() {
     }),
     [copy]
   );
-
-  function normalizeContractTitle(title: string, stableId: string) {
-    if (!isKoLocale) {
-      return title;
-    }
-    const normalized = title.trim();
-    if (normalized.length === 0) {
-      return `계약서 ${stableId.slice(0, 8)}`;
-    }
-    if (/[\uac00-\ud7a3]/.test(normalized)) {
-      return normalized;
-    }
-    const asciiCount = (normalized.match(/[A-Za-z0-9]/g) ?? []).length;
-    if (asciiCount / normalized.length >= 0.6) {
-      return `계약서 ${stableId.slice(0, 8)}`;
-    }
-    return normalized;
-  }
-
   const reload = useCallback(async () => {
     setError(null);
     const [templateBodyRaw, documentBodyRaw] = await Promise.all([
@@ -91,7 +69,6 @@ export default function AdminContractsWorkspace() {
     setTemplates((templateBodyRaw as { templates?: ContractTemplate[] }).templates ?? []);
     setDocuments((documentBodyRaw as { documents?: ContractDocument[] }).documents ?? []);
   }, [copy.loadError]);
-
   useEffect(() => {
     reload().catch((loadError) => {
       setError(
@@ -101,7 +78,6 @@ export default function AdminContractsWorkspace() {
       );
     });
   }, [copy.loadError, reload]);
-
   async function submitTemplate() {
     setError(null);
     setMessage(null);
@@ -126,7 +102,6 @@ export default function AdminContractsWorkspace() {
       );
     }
   }
-
   async function createDraftDocument() {
     if (!selectedTemplateId || employeeId.trim().length === 0) {
       setError(copy.requiredTemplateAndEmployeeError);
@@ -156,7 +131,6 @@ export default function AdminContractsWorkspace() {
       );
     }
   }
-
   async function runDocumentAction(documentId: string, action: ContractDocumentAction) {
     setError(null);
     setMessage(null);
@@ -180,10 +154,9 @@ export default function AdminContractsWorkspace() {
               `${copy.actionFailedPrefix}: ${actionLabelByAction[action]}`
             )
           : `${copy.actionFailedPrefix}: ${actionLabelByAction[action]}`
-      );
+        );
     }
   }
-
   return (
     <main className="saas-content">
       <header className="page-header">
@@ -198,10 +171,8 @@ export default function AdminContractsWorkspace() {
           </div>
         </div>
       </header>
-
       {error ? <p className="inline-error">{error}</p> : null}
       {message ? <p className="small">{message}</p> : null}
-
       <section className="kpi-strip" aria-label={copy.summaryKpiAria}>
         <article className="kpi-card">
           <span>{copy.templatesKpiLabel}</span>
@@ -216,7 +187,6 @@ export default function AdminContractsWorkspace() {
           <strong>{documents.filter((item) => item.approvalStatus === "PENDING").length}</strong>
         </article>
       </section>
-
       <section className="panel-grid">
         <article id="contract-template-library" className="panel panel-contract-template-library">
           <h2>{copy.templateLibraryTitle}</h2>
@@ -251,7 +221,7 @@ export default function AdminContractsWorkspace() {
             {templates.map((template) => (
               <li key={template.id} className={`tone-${template.status === "ACTIVE" ? "ready" : template.status === "DRAFT" ? "watch" : "risk"}`}>
                 <div className="contract-template-head">
-                  <strong>{normalizeContractTitle(template.name, template.id)}</strong>
+                  <strong>{normalizeContractsEntityTitle(template.name, template.id, isKoLocale)}</strong>
                   <span className="queue-history-chip">v{template.version}</span>
                 </div>
                 <div className="contract-template-meta">
@@ -266,7 +236,6 @@ export default function AdminContractsWorkspace() {
             ))}
           </ul>
         </article>
-
         <article id="contract-signature-readiness" className="panel panel-contract-signature-readiness">
           <h2>{copy.documentLifecycleTitle}</h2>
           <div className="contract-form-grid">
@@ -292,7 +261,7 @@ export default function AdminContractsWorkspace() {
             {documents.map((document) => (
               <li key={document.id} className={`tone-${document.status === "SIGNED" ? "ready" : document.status === "REJECTED" ? "risk" : "watch"}`}>
                 <div className="contract-signature-readiness-head">
-                  <strong>{normalizeContractTitle(document.title, document.id)}</strong>
+                  <strong>{normalizeContractsEntityTitle(document.title, document.id, isKoLocale)}</strong>
                   <span className="queue-history-chip">{documentStatusLabels[document.status]}</span>
                 </div>
                 <p>
